[Rename] 컴포넌트 디렉토리 구조 세분화
