+++ components/layouts/sidebar/collected-card.tsx
@@ -3,7 +3,7 @@ import { Card } from '@mui/material';
 import {
   getOverThreeRatingStyle,
   getOverFiveRatingStyle,
-  getRatingStarsStyle,
+  getRatingCircleStyle,
   getCollectedCardStyle,
   getUniqueCardStyle,
   getUniqueCardEffectStyle,
@@ -45,7 +45,7 @@ export default function CollectedCard({
           {Array(ratings)
             .fill(0)
             .map((_, index) => (
-              <div key={index} className={getRatingStarsStyle()}>
+              <div key={index} className={getRatingCircleStyle()}>
                 <i className="fa-solid fa-star absolute text-yellow-300 text-xs"></i>
               </div>
             ))}
