+++ components/layouts/sidebar/footer/tier-badge.tsx
