+++ components/layouts/sidebar/footer/tier-info-modal.tsx
