+++ components/layouts/sidebar/normal-card.tsx
