+++ components/layouts/sub-sidebar/normal/collected-badge.tsx
